+++ src/services/media/__init__.py
@@ -9,9 +9,9 @@ from services.media.backend import MediaBackend, PlaybackState, MediaStatus
 from services.media.backend_factory import create_backend, get_backend_info
 
 __all__ = [
-    'MediaBackend',
-    'PlaybackState',
-    'MediaStatus',
-    'create_backend',
-    'get_backend_info',
+    "MediaBackend",
+    "PlaybackState",
+    "MediaStatus",
+    "create_backend",
+    "get_backend_info",
 ]